RAG/requirements.txt
# C:\RAG\requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
chromadb>=1.3.0
sentence-transformers>=3.0.0
rank-bm25>=0.2.2
PyYAML>=6.0.0
requests>=2.32.0
urllib3>=2.2.0
tqdm>=4.66.0
rich>=13.7.1
schedule>=1.2.0
PyPDF2>=3.0.0
pdfminer.six>=20231228
beautifulsoup4>=4.12.0
lxml>=4.9.0
ftfy>=6.1.0
playwright>=1.44.0
numpy>=2.1.0
